checkpoint for deposit saving

+++ FirstSparrow.Application/Repositories/Abstractions/IMerkleNodeRepository.cs
@@ -5,7 +5,7 @@ namespace FirstSparrow.Application.Repositories.Abstractions;
 
 public interface IMerkleNodeRepository : IBaseRepository<MerkleNode, int>
 {
-    Task<MerkleNode?> GetByNodeCoordinate(uint index, int layer, bool ensureExists, CancellationToken cancellationToken = default);
+    Task<MerkleNode?> GetByNodeCoordinate(long index, int layer, bool ensureExists, CancellationToken cancellationToken = default);
 
     Task<MerkleNode> GetNeighbour(MerkleNode merkleNode, CancellationToken cancellationToken = default);
 }
